Bulk commit
Update EventData final
Add powerupActivated-notify to MP Events
Add eventTimingOut-notify to MP Events
Fix eventEntrantResult-notify on MP Events
Add SoapboxAPI
Add getPersonaData to SoapboxAPI
Add getEventData to SoapboxAPI
Change LobbyTimerStartTime to 30 seconds
Fix minor bugs

+++ src/main/java/br/com/soapboxrace/bo/DriverPersonaBO.java
@@ -46,7 +46,7 @@ public class DriverPersonaBO {
 		profileDataType.setName(personaEntity.getName());
 		profileDataType.setCash(personaEntity.getCash());
 		profileDataType.setIconIndex(personaEntity.getIconIndex());
-		profileDataType.setPersonaId(personaEntity.getId().intValue());
+		profileDataType.setPersonaId(personaEntity.getId());
 		profileDataType.setLevel(personaEntity.getLevel());
 		return profileDataType;
 	}
